requirements.txt
flask
pillow
torch
transformers
ddgs
colorama
